requirements.txt
flask==3.1.2
python-dotenv==1.2.1
google-genai==1.53.0
pytest==9.0.2
